+++ src/runtime/components/base/types/Static.vue
@@ -55,7 +55,19 @@
 </template>
 
 <script setup>
-import { getGlobalTypeData, useI18n, ref, computed, onMounted ,watch, importComp, formatName, formatTexts, showToast } from '#imports'
+import {
+  getGlobalTypeData,
+  useI18n,
+  ref,
+  computed,
+  onMounted,
+  watch,
+  importComp,
+  formatName,
+  formatTexts,
+  showToast,
+  nextTick
+} from '#imports'
 
 const { t } = useI18n();
 
@@ -176,16 +188,18 @@ async function getGlobalType() {
 }
 
 watch(viewSaved, async (val) => {
-  setTimeout(() => {
+  await nextTick()
+  setTimeout(async () => {
     if (val?.dynamicFormRef && props.savedView?.settings) {
+      await nextTick()
       val.dynamicFormRef.settings = props.savedView.settings
     }
-  }, 500)
+  }, 600)
   setTimeout(() => {
     if (val?.dynamicFormRef && props.savedView?.private_data) {
       val.dynamicFormRef.private_data = props.savedView.private_data
     }
-  }, 500)
+  }, 600)
 })
 
 // Lifecycle Hooks